Refactors PushConfig
