+++ pyshgp/push/__init__.py
@@ -1,7 +1,8 @@
 """pyshgp.push"""
 
 from pyshgp.push.instruction_set import InstructionSet
-from pyshgp.push.interpreter import PushConfig, ProgramSignature, Program, PushInterpreter
+from pyshgp.push.interpreter import ProgramSignature, Program, PushInterpreter
+from pyshgp.push.config import PushConfig
 
 __all__ = [
     "PushConfig",
